Better handling of autointegration error messages.
Autointegration will inevitably fail sometimes, for example, when there is not enough light to determine the best integration time. It is important that information about how and why it failed is transmitted to the user and recorded in the log files. The PiccoloSpectrometer module now handles these errors so that they can be logged or displayed as error messages (by other modules).

Currently, the three errors are:

1. No light
2. Unstable light
3. Maximum integration time exceeded